+++ kernel/src/library/hardware/architectures/arch_x86_64/boot.rs
@@ -1,0 +1,9 @@
+impl core::convert::From<&'static mut bootloader::BootInfo> for crate::library::BootInformation
+{
+	fn from(bootloader_information: &'static mut bootloader::BootInfo) -> Self
+	{
+		Self {
+			bootloader_information,
+		}
+	}
+}