+++ kernel/src/library/helper/miscellaneous.rs
@@ -23,6 +23,43 @@ const VERSION: Option<&str> = option_env!("VERSION");
 /// format, i.e. firmware, kernels, etc.
 const TARGET: Option<&str> = option_env!("KERNEL_BUILD_TARGET");
 
+/// ### Kernel `main()` Function
+///
+/// This function is the architecture independent entrypoint for the
+/// kernel.
+///
+/// This function initializes the whole kernel. It takes care of
+///
+/// - printing important initial information
+/// - calling the hardware initialization subroutine
+pub fn main(boot_information: &super::BootInformation) -> !
+{
+	use super::super::{
+		hardware,
+		log,
+		memory,
+	};
+
+	log::set_log_level(log::Level::Trace);
+	display_initial_information(boot_information);
+
+	crate::log_info!("Kernel initialization started");
+
+	hardware::init();
+	memory::init(boot_information);
+
+	crate::log_info!("Kernel initialization finished");
+
+	// It is fine to have this here. If unit tests are run for `lib.rs`,
+	// this will make sure all is initialized beforehand. This does not
+	// affect `main.rs` since we use `crate::` which refers to `lib.rs`
+	// and is therefore not called when running unit tests for `main.rs`.
+	#[cfg(test)]
+	crate::__test_runner();
+
+	never_return()
+}
+
 /// ### The Event Horizon
 ///
 /// This function is just a nice abstraction of the call to `loop
@@ -35,6 +72,7 @@ const TARGET: Option<&str> = option_env!("KERNEL_BUILD_TARGET");
 pub fn never_return() -> !
 {
 	loop {
+		#[cfg(target_arch = "x86_64")]
 		x86_64::instructions::hlt();
 	}
 }
@@ -47,18 +85,12 @@ pub fn never_return() -> !
 /// - the bootloader
 ///
 /// on the serial interface with this function.
-pub(in super::super) fn display_initial_information(boot_information: &bootloader::BootInfo)
+pub(super) fn display_initial_information(boot_information: &super::BootInformation)
 {
 	crate::log!("This is unCORE {}\n", VERSION.unwrap_or("testing"));
 	crate::log_info!(
 		"Target triple reads '{}'",
 		TARGET.unwrap_or("x86_64-unknown-none (defaulted)")
-	);
-	crate::log_info!(
-		"Bootloader version {}.{}.{}",
-		boot_information.version_major,
-		boot_information.version_minor,
-		boot_information.version_patch
 	);
 
 	crate::log_trace!(