+++ kernel/src/library/memory.rs
@@ -4,8 +4,8 @@
 /// uses the `init` function provided by the `hardware` module to use
 /// conditional compilation to initialize the memory correctly for
 /// each target platform.
-pub fn init(boot_information: &bootloader::BootInfo)
+pub fn init(boot_information: &crate::library::BootInformation)
 {
 	crate::log_info!("Initializing virtual memory");
-	crate::library::hardware::memory::init(boot_information);
+	crate::library::hardware::memory::init(boot_information.bootloader_information);
 }